Release artifacts with goreleaser
This commit changes slightly the way the app was built and released.
Build pipeline is tasked to only build and test the release on
PRs/merges to master.

A new pipeline, called `release` is used to prepare a release upon a tag
is created. Goreleaser is used to create tar.gz artifacts and Docker
images (which are pushed to Docker registry)