Drop mmcdole/gofeed dependency
The exporter relied on a third-party library (github.com/mmcdole/gofeed)
to fetch the N0NBH's RSS feed in order to retrieve the data.
N0NBH offers an XML feed as well (under
https://hamqsl.net/solarxml.php).

As we don't rely on more advanced RSS functionality, this commit
replaces gofeed library with a more simplistic HTTP client that performs
the request to retrieve the XML feed instead and parsing. Some
adjustments have been made on XML parsing to support this.

Lastly, this commit adjusts the testdata accordingly.