Perform data collection asyncronously
Existing exporter implementation was acting more or less as a reflector;
Each time a request to `/metrics` was made, Prometheus collector was
syncronously collecting the data from the source. Even though this was
seem straightforward at first, it comes with a set of drawbacks (slight
delay during scrapes, potential data misses if the server experiences
transient errors etc.)

This commit introduces asynchronous data collection from the source as
an improvement. Data are collected asynchronously every 60 seconds
(using a separate goroutine). Each time a request to scrape metrics is
performed, it is served with a copy of the last collected data (which
might be slightly stale). As metrics from N0NBH page are updated every
few minutes, we forsee that serving "stale" copy of data is not an
actual problem atm.